requirements.txt
flask
python-dotenv
flask-cors
openai==0.28